Add fallback behaviour when responding w/ text
The function `add_text` was able to assume the "text" dictionary was
already available because of the way `new_response` was defined.

We no longer assume the response will have text so we start by looking
for the array with text responses and if we don't find it, we then
create it.